+++ changelog.md
@@ -30,6 +30,8 @@ Codion Change Log
 - EditComponentFactory type parameters reordered.
 - EntityTablePanel.userPreferencesKey() renamed preferencesKey(), EntityPanel.preferencesKey() added.
 - EntityEditPanel.writePreference() and applyPreferences() added along with preferencesKey().
+### is.codion.framework.server
+- LocalConnectionHandler.TRACER logger added, handling client method traces.
 
 ## 0.18.45
 ### is.codion.common.core